patchkernel: optimize deletion of interfaces
Deleting the last interface requires some additional work. If the ids
of interfaces to be deleted contain the last interface, that interface
is deleted after deleting all other interfaces. In this way we make
sure to delete the last interface just once (after deleting the last
interface, another interface becomes the last one and that interface
may be on the deletion list as well, and on and so forth).